chore: revert sensor refactor
